+++ packages/ui/dialog/src/InfoDialog/InfoDialogFooter/InfoDialogFooter.tsx
@@ -1,5 +1,5 @@
 import { Button } from '@pikas-ui/button'
-import { styled } from '@pikas-ui/styles'
+import { Colors, styled } from '@pikas-ui/styles'
 
 const Container = styled('div', {
   display: 'flex',
@@ -11,6 +11,7 @@ const Container = styled('div', {
 export interface InfoDialogFooterProps {
   onClose?: () => void
   validateButtonLabel?: string
+  validateButtonColor?: Colors
   onValidated?: () => void
 }
 
@@ -18,11 +19,12 @@ export const InfoDialogFooter: React.FC<InfoDialogFooterProps> = ({
   onClose,
   onValidated,
   validateButtonLabel,
+  validateButtonColor,
 }) => {
   return (
     <Container>
       <Button
-        color="PRIMARY"
+        color={validateButtonColor}
         onClick={(): void => {
           onValidated?.()
           onClose?.()