+++ packages/ui/dialog/src/SuccessDialog/SuccessDialogFooter/SuccessDialogFooter.tsx
@@ -1,5 +1,5 @@
 import { Button } from '@pikas-ui/button'
-import { styled } from '@pikas-ui/styles'
+import { Colors, styled } from '@pikas-ui/styles'
 
 const Container = styled('div', {
   display: 'flex',
@@ -10,16 +10,18 @@ const Container = styled('div', {
 
 export interface SuccessDialogFooterProps {
   validateButtonLabel?: string
+  validateButtonColor?: Colors
   onClose?: () => void
 }
 
 export const SuccessDialogFooter: React.FC<SuccessDialogFooterProps> = ({
   validateButtonLabel,
+  validateButtonColor,
   onClose,
 }) => {
   return (
     <Container>
-      <Button color="SUCCESS" onClick={onClose} width="auto">
+      <Button color={validateButtonColor} onClick={onClose} width="auto">
         {validateButtonLabel}
       </Button>
     </Container>