+++ packages/ui/alert/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pikas-ui/alert
 
+## 3.2.0
+
+### Minor Changes
+
+- ERROR colors to DANGER colors
+
+### Patch Changes
+
+- Updated dependencies
+  - @pikas-ui/styles@3.2.0
+  - @pikas-ui/icons@3.1.2
+
 ## 3.1.1
 
 ### Patch Changes