+++ packages/ui/badge/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pikas-ui/badge
 
+## 1.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @pikas-ui/styles@3.2.0
+  - @pikas-ui/icons@3.1.2
+
 ## 1.1.2
 
 ### Patch Changes