+++ packages/ui/dialog/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @pikas-ui/dialog
 
+## 2.2.0
+
+### Minor Changes
+
+- ERROR colors to DANGER colors
+
+### Patch Changes
+
+- Add button colors props
+- Updated dependencies
+  - @pikas-ui/styles@3.2.0
+  - @pikas-ui/button@3.1.2
+  - @pikas-ui/icons@3.1.2
+  - @pikas-ui/title@1.1.2
+
 ## 2.1.3
 
 ### Patch Changes