+++ packages/ui/dropdown-menu/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @pikas-ui/dropdown-menu
 
+## 3.2.0
+
+### Minor Changes
+
+- ERROR colors to DANGER colors
+
+### Patch Changes
+
+- Update type
+- Export type
+- Updated dependencies
+- Updated dependencies
+  - @pikas-ui/menu@3.2.0
+  - @pikas-ui/styles@3.2.0
+  - @pikas-ui/icons@3.1.2
+  - @pikas-ui/loader@3.1.2
+
 ## 3.1.1
 
 ### Patch Changes