+++ packages/ui/progress/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pikas-ui/progress
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @pikas-ui/styles@3.2.0
+  - @pikas-ui/skeleton@2.1.2
+
 ## 1.1.1
 
 ### Patch Changes