+++ packages/ui/switch/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pikas-ui/switch
 
+## 2.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @pikas-ui/styles@3.2.0
+  - @pikas-ui/text@4.1.0
+  - @pikas-ui/icons@3.1.2
+  - @pikas-ui/tooltip@3.1.2
+
 ## 2.1.3
 
 ### Patch Changes