+++ packages/ui/tabs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pikas-ui/tabs
 
+## 2.1.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @pikas-ui/styles@3.2.0
+
 ## 2.1.5
 
 ### Patch Changes